refactor(runtime): rename std module to hosted
- Rename 'std' module to 'hosted' to better reflect its purpose
- Update import in lib.rs to use new module name
